+++ app/app/Http/Controllers/OcenaController.php
@@ -13,37 +13,23 @@ use Illuminate\Support\Facades\Validator;
 
 class OcenaController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function index()
     {
         return new OcenaCollection(Ocena::all());
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function create()
     {
         //
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
+
     public function store(Request $request)
     {
         $validator = Validator::make($request->all(), [
             'datum_i_vreme' => 'required|date',
-            'korisnik' => 'required|numeric|digits_between:1,5',
             'film' => 'required|numeric|digits_between:1,5',
             'ocena' => 'required|numeric|lte:5|gte:1',
             'poruka' => 'required|string|min:20',
@@ -52,10 +38,12 @@ class OcenaController extends Controller
 
         if ($validator->fails())
             return response()->json($validator->errors());
+        if (auth()->user()->isAdmin())
+            return response()->json('Niste ovlašćeni da kreirate nove ocene.');
 
         $ocena = Ocena::create([
             'datum_i_vreme' => $request->datum_i_vreme,
-            'korisnik' => $request->korisnik,
+            'korisnik' => auth()->user()->id,
             'film' => $request->film,
             'ocena' => $request->ocena,
             'poruka' => $request->poruka,
@@ -68,35 +56,19 @@ class OcenaController extends Controller
     }
 
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\Ocena  $ocena
-     * @return \Illuminate\Http\Response
-     */
+
     public function show(Ocena $ocena)
     {
         return new OcenaResource($ocena);
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Models\Ocena  $ocena
-     * @return \Illuminate\Http\Response
-     */
+
     public function edit(Ocena $ocena)
     {
         //
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Ocena  $ocena
-     * @return \Illuminate\Http\Response
-     */
+
     public function update(Request $request, Ocena $ocena)
     {
 
@@ -112,24 +84,33 @@ class OcenaController extends Controller
         if ($validator->fails())
             return response()->json($validator->errors());
 
+        if (auth()->user()->isAdmin())
+            return response()->json('Niste ovlašćeni da ažurirate ocene.');
+
+        if (auth()->user()->id != $ocena->korisnik)
+            return response()->json('Niste ovlašćeni da ažurirate tuđe ocene');
+
         $ocena->datum_i_vreme = $request->datum_i_vreme;
-        $ocena->korisnik = $request->korisnik;
+        $ocena->korisnik = auth()->user()->id;
         $ocena->film = $request->film;
         $ocena->ocena = $request->ocena;
         $ocena->poruka = $request->poruka;
         $ocena->bioskop = $request->bioskop;
 
+        $ocena->save();
+
         return response()->json(['Ocena je uspešno kreirana.', new OcenaResource($ocena)]);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\Ocena  $ocena
-     * @return \Illuminate\Http\Response
-     */
+
     public function destroy(Ocena $ocena)
     {
+        if (auth()->user()->isAdmin())
+            return response()->json('Niste ovlašćeni da obrišete ocenu.');
+
+        if (auth()->user()->id != $ocena->korisnik)
+            return response()->json('Niste ovlašćeni da obrišete tuđu ocenu.');
+
         $ocena->delete();
 
         return response()->json('Ocena je uspešno obrisana.');
